refactor: remove MlirCompilerTypes.h and inline types
Remove unnecessary header file MlirCompilerTypes.h. The types were only
used in MlirCompiler.cpp, so inline them there in an anonymous namespace.

Changes:
- Delete MlirCompilerTypes.h (41 LOC)
- Inline CompilerErrorCode and CompilerError into MlirCompiler.cpp
- Remove includes from MlirCompiler.h and pass_main.cpp

Benefits:
- One less file to maintain
- Types are private to MlirCompiler.cpp (proper encapsulation)
- No duplicate type definitions across files
- Simpler build

Down to 3 source files: pass_main.cpp, MlirCompiler.cpp, MlirCompiler.h
