+++ backend-mlir-compiler/level-1-pass/src/MlirCompiler.cpp
@@ -13,7 +13,27 @@
 #include <glog/logging.h>
 #include <sstream>
 
-using namespace mlir_compiler_local;
+namespace {
+
+// Error codes from morphizen-mlir-compiler C API
+enum CompilerErrorCode {
+    COMPILER_SUCCESS = 0,
+    COMPILER_ERROR_INVALID_INPUT = 1,
+    COMPILER_ERROR_PARSE_FAILED = 2,
+    COMPILER_ERROR_PASS_FAILED = 3,
+    COMPILER_ERROR_TRANSLATION_FAILED = 4,
+    COMPILER_ERROR_LINKING_FAILED = 5,
+    COMPILER_ERROR_IO_ERROR = 6,
+    COMPILER_ERROR_COMPILATION_FAILED = 7,
+    COMPILER_ERROR_VALIDATION_FAILED = 8,
+    COMPILER_ERROR_INTERNAL = 99
+};
+
+// Error struct from morphizen-mlir-compiler C API
+struct CompilerError {
+    CompilerErrorCode code;
+    char message[1024];
+};
 
 namespace hipdnn {
 namespace level1pass {
